move tests to their own folder
also rename examples -> fixtures
